Changed SMearchX.py to PLearch.py
I decided instead of scraping data, it would be easier to use the browser on a user's computer to search for keywords on a specific platform. This is why I renamed the program to PLearch (Platform Search)

The webbrowser module allows the user to open web pages in the default web browser on their system, without needing to interact directly with the browser application.